refactor(tz): extract general timezone primitives into @app/utils/timezone Move resolveZone, zoneOffsetMinutes, epochFromWallClockInZone (plus the partsInZone helper and the IANA alias map) into a new shared util so other tools can reuse them; tz keeps only natural-language parsing and zone-line display. Relevant tests move to src/utils/timezone.test.ts. (PR #206)
